Add supporting Independent Expenditures to candidates
[Closes caciviclab/disclosure-backend-static#95]

In that issue, Suzanne noticed that we feature "Expenditures Opposing
Candidate" on the candidate's detail page, however we don't include
Expenditures supporting candidate.

Not many candidates have independent expenditures, but those that do --
Huber Trenado, Jumoke Hodge, and James Harris (three candidates for the
school board) -- have quite a lot (more than $100k each).

This also makes the section conditionally hidden if there are no
independent expenditures for/against the candidate. I'm not confident
this is always the right move (it might be nice for a user to know that
a candidate is receiving no IE money) but it seems like a reasonable
first step since most candidates have zero IE money.